Remove duplication of calculation in both versions of question 4.
benefit_options is the list of benefits that are affect the benefit cap.

benefit_options, total_benefits and benefit_cap are being duplicated in the two versions
of question 4 (current and future). Moved the calculations to question 3 as you always
land on a version of question 4 from question 3.

I cherry-picked df7657c9780a755daf232f103f5921c4c703c532 (from PR 2542) and
made the following changes:

* Ran the regression tests.

* Updated the checksum data as the regression tests all pass.