Reduce duplication of total_benefit string formatting
The value for total_benefits is needed in 6 outcomes, and is being formatted in each
outcome.

Move where total_benefits is formatted so it only needs to be done once for each flow.

Rename total_benefits that is displayed in the template to total_benefits_amount so that
the dtring formatting does not affected the amount being calculated.