Reduce duplication of total benefit amount that's over the cap
The total amount was being calculated in the outcome. I've moved this to the previous
question because this value is calculated in the same way regardless of the flow
or whether you are affected by the London, or outside London cap.

I cherry-picked 4a6f66ef875f85ca34f12fcb55ee417974ca7dc3 (from PR 2542) and
made the following changes:

* Updated the checksum data as the regression tests are passing.