Fix YearRange#ends_on for year beginning on leap day (29 Feb)
In this Trello card [1], @erikse pointed out that a year range beginning on 29th
February should end on 28th February the following year and not 27th February as
in the previous implementation.

Also a year range should always include either 365 or 366 days depending on
whether it includes a leap day.

This commit adds more test coverage to highlight the problem as well as fixing
the implementation.

[1]: https://trello.com/c/M17mqlSN