Make validation more consistent in marriage-abroad
By extracting a new method, `MarriageAbroadCalculator#valid_ceremony_country?`,
and calling that method from a `validate` block within the flow definition, we
can make the code more consistent with the approach we've adopted elsewhere.

This also has the benefit of not explicitly raising a
`SmartAnswer::InvalidResponse` exception from within the
`MarriageAbroadCalculator` as per this Trello card [1].

All the `marriage-abroad` regression tests still pass without any changes to the
artefacts and so I'm happy to update the checksums accordingly.

[1]: https://trello.com/c/aPkzlCZb