Remove unused option from MarriedCouplesAllowanceCalculator
Looking at this commit [1] it appears as if the `validate_income` option was
introduced in order to retain some behaviour in a v1 flow when a new v2 flow was
being introduced. This v2 flow then replaced the v1 flow in this commit [2] at
which point the calculator was only ever instantiated with this option set to
`false` (apart from in a unit test).

This commit removes the `validate_income` option altogether which simplifies the
code and avoids raising a `SmartAnswer::InvalidResponse` exception as per this
Trello card [3].

The regression tests for `calculate-married-couples-allowance` all pass without
any changes to the artefacts and so I've updated the regression test checksums
accordingly.

[1]: 646cace9a136e3a900008a05a9bb81a24d7cc97f
[2]: 237dfe978b9b0b8c19e70cedbb1a8472aee4e8bb
[3]: https://trello.com/c/aPkzlCZb