Simplify StatePensionTopupCalculator#too_young?
There's no need for the `else` statement in this method, because the value of
`gender` is always either the default value ('female') or the response from the
`gender?` multiple choice question. The latter has some built-in validation [1]
which prevents a response other than 'female' or 'male'. I've tested this by
hacking the URL for this question in production [2] and seeing that 'foo' is not
allowed as a response.

This also has the advantage of avoiding raising a `SmartAnswer::InvalidResponse`
exception as per this Trello card [3].

Since all the `state-pension-topup` regression tests pass without any changes to
the artefacts, I've updated the regression test checksums accordingly.

[1]: https://github.com/alphagov/smart-answers/blob/2bb62d796b6191b9dcfda2d3fe923658afb437ee/lib/smart_answer/question/multiple_choice.rb#L24
[2]: https://www.gov.uk/state-pension-topup/y/1940-05-11/foo
[3]: https://trello.com/c/aPkzlCZb