Add Flow#append to append nodes from one flow to another
My intention is to use this method to replace the use of `use_shared_logic`
which includes the undesirable use of `Kernel#eval`.

Unfortunately it's necessary to add the `flow` attribute *writer* so that the
shared logic can use node templates in the directory for the flow which includes
it. This is because `Node#template_directory` uses the value of `@flow` which is
usually set in `Node#initialize` and in this case it would be an instance of the
shared logic flow and not the including flow. I don't really like having to add
this attribute writer method, but I don't think it's too bad.

Note that at this point the `Flow#use_shared_logic` method still exists. I plan
to gradually phase out it's use in subsequent commits until I can remove it
entirely.