WIP: Introduce Flow#add method to add one flow to another
Adding `Node#flow=` is currently necessary if we want the shared logic to use
node templates in the directory for the flow which includes it. This is because
`Node#template_directory` uses the value of `@flow` which is usually set in
`Node#initialize` and in this case it would be an instance of the shared logic
flow and not the including flow. I don't really like having to add this
attribute writer method, but I haven't yet come up with an alternative that I
prefer.

TODO:

[ ] Try to come up with an alternative implemenation which addresses the point
    above.
[ ] Add tests for new behaviour