Fix YearRange#ends_on for year ending on leap day (29 Feb)
In implementing the fix for a year beginning on a leap day [1], I realised that
there was a similar issue for a year ending on a leap day. In both of these
edge cases it seems somewhat subjective whether or not the leap day should be
included. It's my sense that the most intuitive behaviour is to always include
the leap day. However, it's possible that specific policies/legislation might
have particular rules about this. We should bear this in mind when we make use
of the `YearRange` class and potentially make this behaviour configurable if
necessary.

This commit adds more test coverage to highlight the problem as well as fixing
the implementation.

[1]: https://trello.com/c/M17mqlSN