Revert "Remove Outcome#transition method"
This reverts commit 0929036ffa0c92dbea372c2d570fe6f25a087e29.

It was a mistake to remove the `Outcome#transition` method which raised a
`SmartAnswer::InvalidNode` exception. I hadn't noticed that this exception is
rescued and the app responds with a "404 Not Found" [1]. The latter seems like
sensible behaviour and so I'm restoring the previous implementation in this
commit. This should prevent exceptions like this one [2] from appearing in
Errbit.

[1]: https://github.com/alphagov/smart-answers/blob/bf77a584737d8dbfee5694b64d5ad8870ebe0445/app/controllers/smart_answers_controller.rb#L7
[2]: https://errbit.publishing.service.gov.uk/apps/533c35ae0da1159384044f5f/problems/572087f96578631e0c660b00