Extract payment info from outcome_opposite_sex_marriage_in_japan
I updated and ran the add-payment-partials-to-marriage-abroad-services-data.rb
script to update marriage_abroad_services.yml with the payment partial for
Japan. There was no output from the script and I can see that the
marriage_abroad_services.yml file has been updated as expected.

I haven't run the regression tests in this commit but am confident that there
won't be any changes. I will run the regression tests once I've made all the
changes in this branch to catch anything that I might've missed.