Extract payment info from outcome_same_sex_civil_partnership_in_affirmation_countries
I used the following command to determine the countries that lead to this
outcome, so that I can be sure I'm adding payment information for all necessary
countries:

    $ rails r script/marriage-abroad-countries-leading-to-outcome.rb \
    outcome_same_sex_civil_partnership_in_affirmation_countries

I've updated the add-payment-partials-to-marriage-abroad-services-data.rb
script based on the countries that lead to, and the logic in, this outcome.

Running add-payment-partials-to-marriage-abroad-services-data.rb results in no
output. There are no changes to marriage_abroad_services.yml which means that
the payment information for both these countries already exists in that file.

I haven't run the regression tests in this commit but am confident that there
won't be any changes. I will run the regression tests once I've made all the
changes in this branch to catch anything that I might've missed.