Extract payment info from outcome_same_sex_civil_partnership
I used the following command to determine the countries that lead to this
outcome, so that I can be sure I'm adding payment information for all necessary
countries:

    $ rails r script/marriage-abroad-countries-leading-to-outcome.rb \
    outcome_same_sex_civil_partnership

I've updated the add-payment-partials-to-marriage-abroad-services-data.rb
script based on the countries that lead to, and the logic in, this outcome.
Note the excluded countries in this script: The Czech Republic and Côte
d’Ivoire are excluded to match the behaviour in the outcome template.

Running add-payment-partials-to-marriage-abroad-services-data.rb results in:

Entry not found for brazil.
Payment method partial already set for finland. Existing value: pay_in_euros_or_visa_electron. New value: pay_by_cash_or_credit_card_no_cheque.

I've updated the regression test artefacts to see the effect of these changes.

The "Entry not found for ..." countries will no longer display payment
information when reaching this outcome. This is correct as the fact that they
don't have an entry in marriage_abroad_services.yml means that they shouldn't
show any payment information. This change can be seen in the regression test
artefacts for Brazil.

The regression test artefact changes for Finland are the result of previously
inconsistent payment information being displayed for Finland. Storing the
payment information partial in marriage_abroad_services.yml means that we
always show the same payment information for a single country. I've asked what
the correct payment information is for Finland.