Extract payment info from _fees.govspeak.erb
The `outcome_opposite_sex_marriage_in_affirmation_countries` outcome is the
only outcome that renders the `fees` partial. I used the following command to
determine the countries that lead to this outcome:

    $ rails r script/marriage-abroad-countries-leading-to-outcome.rb \
    outcome_opposite_sex_marriage_in_affirmation_countries

I created the add-payment-partials-to-marriage-abroad-services-data.rb to
update the data in marriage_abroad_services.yml. I plan to update this script
as I continue to extract the payment info from the templates.

Running the script in this commit results in no output which means that the
payment information for each of the countries listed has been successfully
added to marriage_abroad_services.yml.