Extract payment info from _fees.govspeak.erb
I'm storing the name of the payment information partial to display in
marriage_abroad_services.yml, rather than relying on logic in the template.
This will make it easier to update services and payment information for
specific countries as we won't have to untangle the logic in the flow/templates
to work out what's being displayed.

The _fees.govspeak.erb partial is only rendered by
`outcome_opposite_sex_marriage_in_affirmation_countries`.  I used the following
command to determine the countries that lead to this outcome, so that I can be
sure I'm adding payment information for all necessary countries:

    $ rails r script/marriage-abroad-countries-leading-to-outcome.rb \
    outcome_opposite_sex_marriage_in_affirmation_countries

I then created the add-payment-partials-to-marriage-abroad-services-data.rb
script to programatically update the data in marriage_abroad_services.yml based
on the countries that lead to
`outcome_opposite_sex_marriage_in_affirmation_countries` and the logic in
_fees.govspeak.erb. I plan to update the data in this new script as I continue
to extract the payment info from the templates.

Running add-payment-partials-to-marriage-abroad-services-data.rb results in no
output which means that the payment information for each of the countries in
the array has been successfully added to marriage_abroad_services.yml.

The marriage-abroad regression tests are passing in this commit.