Remove payment info from outcome_opposite_sex_in_no_cni_countries_when_residing_in_ceremony_or_third_country
This outcome doesn't display the _consular_fees_table partial so shouldn't
render the payment information partial either.

NOTE. The link_to_consular_fees_pay_in_local_currency partial should almost
certainly be removed too but I'll deal with that separately.

I've confirmed that the changes to the regression test artefacts look correct.
I compared the list of countries that lead to this outcome to the changes to
the artefacts:

I used the following command to get the list of countries leading to this
outcome:

    $ rails r script/marriage-abroad-countries-leading-to-outcome.rb \
    outcome_opposite_sex_in_no_cni_countries_when_residing_in_ceremony_or_third_country

NOTE. The remainder of this commit note was written when I had a complete set
(i.e. for every ceremony country) of regression test artefacts for
marriage-abroad. I think the information is still useful so I'm leaving it in.

The following countries lead to this outcome and have had the payment
information removed from their artefacts:

* afghanistan
* american-samoa
* andorra
* benin
* bhutan
* burkina-faso
* cape-verde
* central-african-republic
* chad
* comoros
* congo
* djibouti
* equatorial-guinea
* eritrea
* gabon
* guinea
* guinea-bissau
* haiti
* iraq
* liechtenstein
* mali
* marshall-islands
* mauritania
* micronesia
* niger
* palau
* paraguay
* rwanda
* san-marino
* sao-tome-and-principe
* south-sudan
* suriname
* timor-leste
* togo
* western-sahara

The following countries lead to this outcome but are in the
`COUNTRIES_WITHOUT_CONSULAR_FACILITIES` array (i.e.
`#country_without_consular_facilities?` is true) so they weren't displaying any
payment information anyway.

* aruba
* bonaire-st-eustatius-saba
* burundi
* curacao
* saint-barthelemy
* st-maarten
* st-martin