WIP: Use Mocha instead of WebMock in regression tests
I ran the marriage-abroad regression tests three times and got the following
values for `real` time:

0m43.545s
0m44.086s
0m45.740s

It also results in a number of regression test artefact changes as country
names are different ("American-samoa instead of American Samoa") and embassy
data no longer exists (because I'm stubbing `fco_organisation` to return
`nil`).

I have to move `setup_worldwide_locations` so that it's run every time `setup`
is run. If I leave it where it was then I get WebMock failures because some
code is trying to make real requests. I believe this is because the Mocha setup
happens first time round, then it's reset in the `teardown` and never set-up
again.