Introduce configurations for regression tests
I want to make the stubbed current time configurable so that we can
incrementally update this for individual flows.

Although it's not strictly necessary, I've chosen to introduce a mechanism
which will support other configuration options in the future. I think this is
actually easier to understand.

The idea is that every configuration option will have a value in the
default configuration with optional values in flow-specific sections.

Thus with the following `configurations.yml`, one can obtain a value for the
`my_key` option for the current flow as follows:

```
smart_answer_helper = SmartAnswerTestHelper.new(flow_name)
smart_answer_helper.configuration.fetch(:my_key)
```

```
---
  default:
    my_key: 123
  "my-flow":
    my_key: 456
```

When `flow_name` == "my-flow", the value will be `456`, and otherwise it will
be the default value, `123`.