Introduce configurations for regression tests
I want to make the stubbed current time configurable so that we can
incrementally update this for individual flows.

Although it's not strictly necessary, I've chosen to introduce a mechanism
which will support other configuration options in the future. I think this is
actually easier to understand.

The idea is that every configuration option will have a value in the
default configuration with optional values in flow-specific sections.