Make hard-coded regression test current time configurable
I want to make the stubbed current time configurable so that we can
incrementally update this for individual flows.

This value will be used by default, but it can be overridden for individual
flows.

This commit should not result in any change to the stubbed current time for
any of the regression tests.