Update regression test current date for calculate-your-child-maintenance
I've overridden the default value of `current_time` specifically for this flow.

I've chosen a value, `2016-05-05`, which is in the current (2015-2016) tax year.
The previous value, `2015-01-01`, was in the 2014-2015 tax year i.e. two tax
years ago.

This flow does not appear to be date-dependent.

All the question page artefacts have changed, but as usual only because the
"Last updated" value has changed. I screened these out of the diff using the
following command:

    git diff | grep --invert-match '<p class="modified-date">'