Fail fast in regression test if response is not success
When generating regression test artefacts for the first time, we should never
store a response with a non-success status code - this would be a sign that
we've chosen the wrong questions & responses.

When running regression tests on subsequent occasions, it's more useful to see
an assertion failure specifically about the non-success response, rather than an
assertion about the response body diff not matching.

I did consider avoiding the duplication by putting the assertion inside
`SmartAnswerTestHelper#save_output`, but I wanted to use `ActionDispatch::Assertions::ResponseAssertions#assert_response` [1] which
checks for status codes in the range 200-299 and that would've been awkward to
do outside the `SmartAnswersRegressionTest` class.

[1]: http://api.rubyonrails.org/classes/ActionDispatch/Assertions/ResponseAssertions.html#method-i-assert_response