Fix stubbing of current time in regression tests
Since #2518 was merged, when multiple regression tests were run the call to
`Timecop.freeze` in the `setup` block was only being made the first time the
`setup` block was executed. This is because of the optimisation introduced in
this commit [1].

It so happens that #2518 changed the stubbed value of current time in the first
flow, `additional-commodity-code`, to `2016-05-05` i.e. different from the value
used to generate the artefacts for the majority of the flows. So when running
all the regression tests, the stubbed value of current time was set to the value
for `additional-commodity-code` and never changed. This caused the full
regression test suite to fail on the `govuk_smart_answers_regressions` CI
server.

This commit moves the call to `Timecop.freeze` before the condition which skips
the `setup` on subsequent tests. This means that it is now executed for every
test and the appropriate values of current time should be used for each flow.

[1]: c86ba038821012a70cb5d8ee3c462fd6f9c5e53c