Update questions-and-responses for marriage-abroad
I'm in the process of adding the services/fees information to the outcomes for
France, Monaco and the French overseas territories (that is, France, Monaco as
well as countries in `FRENCH_OVERSEAS_TERRITORIES`).

The full set of countries affected is:

* france
* french-guiana
* french-polynesia
* guadeloupe
* martinique
* mayotte
* monaco
* new-caledonia
* reunion
* st-pierre-and-miquelon
* wallis-and-futuna

I've added the missing countries so that we generate regression test artefacts
for all countries listed above.