Display services/fees in outcome_marriage_in_france_or_french_overseas_territory
Si Stephens has confirmed that we don't need to display a link to the consular
fees, or any payment information (e.g. "you can pay by cash") in this outcome.

This outcome is shown for marriages in France, New Caledonia and Wallis and
Futuna (countries that offer PACS) and is shown for all ceremonies in the other
French overseas territories.