Generate questions & responses using production whitehall
This script runs in the Rails `development` environment.

Prior to #2498, being in the `development` environment meant that the app always
used the production instance of the Whitehall app as the backend for the GDS
Worldwide API.

After #2498, by default the app would try to use a local development instance of
the Whitehall app, unless the `PLEK_SERVICE_WHITEHALL_ADMIN_URI` is set.

Since the existing `*-questions-and-responses.yml` files were generated with the
production instance of the Whitehall app, it seems more sensible to set the
`PLEK_SERVICE_WHITEHALL_ADMIN_URI` environment variable to ensure that this
continues to be the case.