Add stub_worldwide_locations test helper method
I'm going to use this method instead of
`GdsApi::TestHelpers::Worldwide#worldwide_api_has_locations` in the integration
tests. This appears to shave about a minute off the time it takes to run `rake`
(i.e. all but the regression tests) on my machine.

I'm committing this first and will commit the change to each integration test
separately.

This works by stubbing `WorldLocation.all` and `WorldLocation.find` to return
some stub `WorldLocation` objects. For each of those objects I've had to stub
`#slug`, `#name` and `#fco_organisation` to ensure that all the tests are
happy.

This is in contrast to `#worldwide_api_has_locations` which works by
configuring WebMock to return JSON data in response to requests to the GOV.UK
worldwide API. Using `#worldwide_api_has_locations` means that we're relying on
the real implementation of `GdsApi::Worldwide` to request the JSON and turn it
into `OpenStruct`'s to be consumed within `WorldLocation.all` and
`WorldLocation.find`.