Fix age calculation - should use end of current tax year
I'm pretty confident that the important age for personal allowance is the age at
the end of the current tax year and not 5th April in the current calendar year
which is how this was implemented in the `PersonalAllowanceCalculator`.

I haven't been able to find any direct (GOV.UK) evidence that this is correct,
but it seems logical and I have found indirect evidence, e.g. this PDF [1].

Ultimately whether or not this is correct isn't that important, because we're
already well past 2013-04-05 which is when the calculation switched to use
date of birth instead of age.

This is also why I haven't written exhaustive test cases - I just tweaked the
Timecop date in the "before 2013 to 2014 tax year" context to force me to fix
the implementation.

[1]: http://www.pwc.co.uk/assets/pdf/budget-2012-tax-datacard-2012.pdf