Introduce DateOfBirth class
Someone's age on a given date depends on whether or not they've had their
birthday in the current year. Also someone who's birthday falls on 29th Feb in a
leap year, is legally considered to have their birthday on 1st March in a
non-leap year [1].

The age calculation in `PersonalAllowanceCalculator#age_on_fifth_april` is not
correct. I plan to use the implementation in this commit to fix this. The same
implementation could probably be used elsewhere,
e.g. in `StatePensionTopupCalculator`.

[1]: https://en.wikipedia.org/wiki/February_29#Legal_status