Add extra highest earner income responses to regression test
This is in preparation for fixing the flow to use the correct income limit for
the current (2016-2017) tax year i.e. £27,700 instead of £27,000.

By including responses just above the old limit and just above the new limit, we
should be able to see the change in behaviour after the fix.

The extra responses were added to the questions & responses file, the responses
& expected results file was then regenerated, and finally the acceptance tests
were run and the new artefact files added along with the updated checksums. All
the regression tests are passing at this point.