Introduce gift_aided_donations state variable
The `calculate_adjusted_net_income` method will return `income` if all of the
following are zero: `gross_pension_contributions`, `net_pension_contributions`,
and `gift_aided_donations`, so it feels simple to use this in *all* scenarios
rather than set `income` earlier in the flow and then override it in some
scenarios.

Although this introduces some duplication in the two "done" outcomes, I'm
planning to remove this in subsequent commits.