Add tests for personal allowance in post-2013-2014 & fix implementation
The age-related personal allowance was only based on age up until the end of the
2012-2013 tax year. Since then it should have been based on the person's date of
birth [1]. This commit introduces tests to check this and fixes the
implementation of `PersonalAllowanceCalculator#age_related_allowance`.

The current date for the regression tests for this flow is 2016-05-05 i.e. in
the 2016-2017 tax year. The Personal Allowance levels for this tax year are all
£11,000 i.e. there is no difference between `personal_allowance`,
`higher_allowance_1` & `higher_allowance_2`. So even though the new
date-of-birth related mechanism is being triggered this doesn't result in any
change in the externally visible behaviour. Hence no changes to the regression
test files are needed.

If we had wanted to see the change in the external behaviour, it would've been
possible to change the current date for the regression tests to be in the
2014-2015 tax year. This was the last tax year in which there was a difference
in the Personal Allowance levels. However, since that tax year is now well past,
I decided that that the unit tests were sufficient.

[1]: https://www.gov.uk/government/publications/rates-and-allowances-income-tax/income-tax-rates-and-allowances-current-and-past#personal-allowances-for-people-born-before-6-april-1948