Remove "duplicate" husband income response from regression test
Now that we're using the correct income limit for the current (2016-2017) tax
year i.e. £27,700 instead of £27,000, the £27,001 response is no longer useful,
because it generates essentially the same artefacts as the £20,000 response.

The £27,001 response was removed from the questions & responses file, the
responses & expected results file was regenerated, and the acceptance tests were
run to generate the artefacts.

All the regression tests are passing at this point, so I've also updated the
checksums.