Fix constant warnings in gds_api/json_client in development env
I recently started seeing the following two warnings for each of the three
constants defined in `gds_api/json_client`:

```
.bundle/gems/gems/gds-api-adapters-25.1.0/lib/gds_api/json_client.rb:65: warning: already initialized constant GdsApi::JsonClient::DEFAULT_TIMEOUT_IN_SECONDS
.bundle-2.3.0/gems/gems/gds-api-adapters-25.1.0/lib/gds_api/json_client.rb:65: warning: previous definition of DEFAULT_TIMEOUT_IN_SECONDS was here
.bundle/gems/gems/gds-api-adapters-25.1.0/lib/gds_api/json_client.rb:66: warning: already initialized constant GdsApi::JsonClient::DEFAULT_CACHE_SIZE
.bundle-2.3.0/gems/gems/gds-api-adapters-25.1.0/lib/gds_api/json_client.rb:66: warning: previous definition of DEFAULT_CACHE_SIZE was here
.bundle/gems/gems/gds-api-adapters-25.1.0/lib/gds_api/json_client.rb:67: warning: already initialized constant GdsApi::JsonClient::DEFAULT_CACHE_TTL
.bundle-2.3.0/gems/gems/gds-api-adapters-25.1.0/lib/gds_api/json_client.rb:67: warning: previous definition of DEFAULT_CACHE_TTL was here
```

I see them when I start the Rails console and when I run the tests with `rake`.

I'm not quite sure why I've only recently started seeing these warnings, but it
might be connected to the fact that I deleted all my bundled gems and
re-installed them.

The fundamental reason we are seeing these warnings is that the
`gds-api-adapters/lib/gds_api/json_client.rb` file is being loaded multiple
times.

One place it's happening is in `lib/working_days.rb` which has the following
two lines at the top:

```
require 'gds_api/base'
require 'gds_api/json_client'
```

The `WorkingDays.load_bank_holidays` method should only need the second of
these lines i.e. `require 'gds_api/json_client'`, so it can use
`GdsApi::JsonClient`. However, it turns out that the latter also depends on
`GdsApi::Base`, but does not require it. This is why we need the first of the
above lines, i.e. `require 'gds_api/base'`.

To further complicate matters, it turns out that `gds_api/base` requires
`gds_api/json_client`! Also all the require statements within the
`gds-api-adapters` gem use `require_relative`.

So I think what's happening is:

```
require 'gds_api/base' (lib/working_days.rb)
  -> require_relative 'json_client' (gds-api-adapters/lib/gds_api/base.rb)
    -> loads 'gds-api-adapters/lib/gds_api/json_client.rb' (1st time)
require 'gds_api/json_client' (lib/working_days.rb)
  -> loads 'gds-api-adapters/lib/gds_api/json_client.rb'   (2nd time)
```

I believe that the usual `require` mechanism which prevents files being loaded
twice is not working, because the paths in each underlying call to `require`
are different, because of the use of `require_relative` in one case and
`require` in the other.

To be fair to `gds-api-adapters`, I don't think `GdsApi::JsonClient` was
intended for use outside the gem. And perhaps the "right" way to fix this is to
extract a `GdsApi::BankHolidays` component from the Smart Answers app into the
gem.

Another way to fix this would be to ensure that the dependency of `JsonClient`
on `Base` is made explicit within the `gds-api-adapters` gem. However, since
we're many versions behind the current version of this gem, it might take us a
while to get a fix to the gem accepted and published and then get the app
updated to use the latest version.

In the short term, in this case we can avoid the warnings by only requiring
`gds_api/base` in `lib/working_days.rb` and relying on that to require
`gds_api/json_client`.

I think this should avoid the warnings at least in the development environment
e.g. when you start a Rails console or server.

Note that I've updated the regression test checksums for
`vat-payment-deadlines`, because that flow depends on the `working_days` file.
The regression tests all still pass, so I'm happy to update the checksums.