Update checksum data for benefit-cap-calculator
This was cherry-picked from 9fe9d7aef885a69e47bf94d037052d1308781a04 and
modified so that it could become the first commit in this branch.

These two additional files should've been added to the checksum data in PR

We're adding these files as they affect the behaviour of the
benefit-cap-calculator flow. We want the regression tests to run when we notice
changes to these files.