Add receiving_non_exemption_benefits_future? question
The options for the `receiving_non_exemption_benefits` questions are stored in
the benefit_cap_data.yml file. Unfortunately, the value of `chosen_cap` isn't
available at question definition time so we have to duplicate this question and
use hardcoded values of `:default` and `:future` in the call to
`config.benefits`.

I manually applied some of 73243a6a40ab9e4ded04aeac0fdfe3913135279b and made
the following changes:

* Updated the integration test to capture the change in behaviour between
  the `default` and `future` flows.

* Added responses for this new question to
  benefit-cap-calculator-questions-and-responses.yml and used those changes to
  regenerated benefit-cap-calculator-responses-and-expected-results.yml.

* Added new question page regression test artefact for the question added in
  this commit.

* Updated the checksum data. The regression tests are passing so I'm happy to
  update this.