Consider the region when getting weekly benefit cap data
Use the GDS Imminence API to get the area that a postcode is in. The responses
are returned as an OpenStruct.

Add methods to determine if the area is in london. Also use the safe navigation
operator (&.) to ensure that a nil is returned in the event something goes
wrong with imminence's response.

I've manually applied some of 9f02fcb0c847407fdc035f25cae60b914b548323 and made
the following changes:

* Modified the commit note slightly, by removing the points that no longer
  apply to this commit.

* Added the London rates to benefit_cap_data.yml. These are identical to
  the current (i.e. pre Autumn 2016 change) National rates. I'll update the rates
  separately.

* Updated the tests based on the changes in
  174d0813933904f23d2e4bfe7852fbc5da36b58d. This removes some duplication by
  using the `imminence_has_areas_for_postcode` helper method instead of using
  `WebMock` directly.

* Updated the integration tests so that they continue to pass now that
  BenefitCapCalculatorConfiguration uses Imminence.

* Updated the smart_answer_regression_test.rb setup so the
  benefit-cap-calculator regression tests pass now that
  BenefitCapCalculatorConfiguration uses Imminence to lookup Postcode data.

* Updated the checksum data as the regression tests are passing.