Update Autumn 2016 benefit cap rates for outside London
I've taken these rates from those added in
5fc48d741d6d96a6f59f28473ae3db184806694e (in PR #2542).

I've updated the benefit-cap-calculator-responses-and-expected-results.yml
file. The changes are due to the rate changes for couples and lone parents
outside of London. The benefit cap has been reduced from £500 to £384.62. The
two changes are for when the user has said they earn £400 in benefits. That
used to be below the cap (hence the previous
`:outcome_not_affected_less_than_cap`) and now it's above (hence the new
`:outcome_affected_greater_than_cap`.

NOTE. I had to point at the production instance of Imminence to update this
file:

    $ PLEK_SERVICE_IMMINENCE_URI=https://imminence.publishing.service.gov.uk \
    rails r script/generate-responses-and-expected-results-for-smart-answer.rb \
    benefit-cap-calculator

If I didn't do that then the script would ultimately connect to
http://imminence.dev.gov.uk (which returns 403 on my machine). The effect of
this was that all post codes would return nil and therefore the
benefit-cap-calculator-responses-and-expected-results file would also contain
changes for the London postcode (because the
`BenefitCapCalculatorConfiguration#weekly_benefit_cap` methods default to
`national` if they can't find the postcode data).

I've run the regression tests and committed the artefacts based on these rate
changes.

* The artefacts that have changed are all for outcome pages reached when using
  the Birmingham (i.e. outside London) postcode.

* The outcome page artefact changes for single people show the change in rates
  from £350 to £257.69 as expected.

* The outcome page artefact changes for couples and lone parents show the
  change in rates from £500 to £384.62 as expected.

I've updated the checksum data as I'm happy with the changes to the regression
test artefacts.