Use production Imminence when generating responses-and-expected-results
In the commit titled "Update Autumn 2016 benefit cap rates for outside London"
I explained that I needed to use the production version of Imminence so the
flow was being exercised in the context of real postcode data. It turns out
that I needed to continue setting the PLEK_SERVICE_IMMINENCE_URI environment
variable everytime I wanted to run the
generate-responses-and-expected-results-for-smart-answer.rb script. Setting the
environment variable in the script itself feels like the correct place.