Update outcome_not_affected_no_housing_benefit_future
I've manually applied the change to this outcome from
573e7f6b0f63b2bc8a1e1b755364b43b0c7da69e (in PR 2542).

I've run the regression tests and committed the changes to the regression test
artefacts. I'm happy that the outcome artefact changes match the change to the
template in this commit.

I've updated the checksum data as the regression tests are passing.