Extract oru_documents_variant_for_birth? method in RegistrationsDataQuery
Since the `oru_result` outcome template in `register-a-birth` already has access
to an instance of this class, it seems better to encapsulate the constant within
the class and introduce this new method.

Note that my immediate motivation for doing this was to avoid a problem when
generating indexable content i.e. `MethodMissingObject` does not define the
`ORU_DOCUMENTS_VARIANT_COUNTRIES_BIRTH` constant.

I've updated the regression test checksums for the relevant flows, because this
change does not alter the behaviour of any of them.