Revert "Handle template exceptions when generating indexable_content"
This (mostly) reverts commit 50ab9b8d4963530eae7142f230d07ae75d04fb93.

Note, however,  that I have *not* reverted the introduction of
`IndexableContentTest`, because it's still a useful test to make sure we don't
introduce any more code which results in raising exceptions.

This commit was originally introduced so that we could generate
`indexable_content` for the `outcome_os_other_countries` outcome in
`marriage-abroad`. We had to add this special case to handle the potential
exception raised by the outcome.

This later pull request [1] updates the `outcome_os_other_countries` outcome
such that it no longer raises an exception, which means that we can simplify the
code by reverting this commit.

[1]: #2504