Extract oru_courier_by_high_commission? method in RegistrationsDataQuery
Since the `oru_result` outcome templates in `register-a-birth` &
`register-a-death` already have access to an instance of this class, it seems
better to encapsulate the constant within the class and introduce this new
method.

Note that my immediate motivation for doing this was to avoid a problem when
generating indexable content i.e. `MethodMissingObject` does not define the
`ORU_COURIER_BY_HIGH_COMISSION` constant.

I've updated the regression test checksums for the relevant flows, because this
change does not alter the behaviour of any of them.