Test LandlordImmigrationCheckCalculator#areas_for_postcode with invalid postcode
Previously this test was just testing `GdsApi::Imminence#areas_for_postcode`.
While the test was useful from an illustrative point of view, I don't think it
belongs in this test case which is supposed to be testing the
`LandlordImmigrationCheckCalculator` class.

I think it's more useful to test
`LandlordImmigrationCheckCalculator#areas_for_postcode` and so I've made that
method public and converted the test to do just that.