LandlordImmigrationCheckCalculator#areas_for_postcode should return array
If the ternary condition is false, then we should return an empty array and not
an empty hash, because the former is what clients of this method are expecting.
e.g. `LandlordImmigrationCheckCalculator#postcode_within?`.

Note that the reason why the unit tests were all passing was because the ternary
condition is never false (even when the postcode is unknown), but I wanted to
fix the code before removing it in a later commit, so that the git history makes
more sense.