Improve name of tests in LandlordImmigrationCheckCalculatorTest
I think the first of these scenarios is better explained as having an "unknown"
postcode rather than an "invalid" one. It strikes me that "E15" is at least in
some way a "valid" postcode, but it does not exist - hence the 404.

Similarly, I don't think it's particularly useful to describe the other
postcodes as "valid".