Refactor LandlordImmigrationCheckCalculator#rules_apply?
I don't think the `postcode_within?` method was a useful abstraction. I think
`countries_for_postcode` is a more useful abstraction.

Although the rules may at some point apply in other countries within the UK,
while they only apply in England, I think it's better not to worry about
multiple countries and keep the code simpler.

Note that this commit highlighted to me the fact that there may be edge cases
where a postcode is in multiple areas, some of which are in England and some of
which are not. This may be possible, because the areas are of varying sizes. At
the moment the code will consider a postcode to be in England if any of its
areas are in England.

This commit does not change the behaviour at all, but I thought it was worth
mentioning. An obvious way to fix this would be to only consider areas of the
country type ("CTR"), however, it doesn't look as if Imminence returns such
areas.