Update plek from v1.7.0 to v1.12.0
I want to update gds-api-adapters to the latest version (v30.6.0) and that
depends on the plek version being >= v1.9.0.

I looked through CHANGELOG for plek and didn't see anything that worried me.

I made the change to version of plek in the Gemfile and ran
`bundle update plek`. I then inspected the changes to Gemfile.lock
and since only the version of plek was affected, I accepted the changes as they
stood.

I've also run all the non-regression tests and they all pass. I intend to run
all the regression tests locally before I submit this branch as a PR.

Note that I've stuck with constraining plek to a specific version in the
Gemfile, even though I think this is overkill. We seem to be inconsistent about
how we approach the constraints on gems in the Gemfile, but I'm not going to
worry about that in this PR.