Update gds-api-adapters from v25.1.0 to v30.6.0
I noticed that we were using quite an old version of gds-api-adapters and I
thought it would be sensible to get more up-to-date.

Also I'm contemplating pushing some functionality out of the smart-answers app
and into gds-api-adapters. Being up-to-date with gds-api-adapters will make this
easier.

I looked through CHANGELOG for gds-api-adapters and didn't see anything that
worried me. Most of the changes seemed to be additions to the API.

I made the change to version of gds-api-adapters in the Gemfile and ran
`bundle update gds-api-adapters`. I then inspected the changes to Gemfile.lock
and removed the following unnecessary changes:

* domain_name (v0.5.25 -> v0.5.20160310)
* mime-types (v2.99.1 -> v2.99.2)
* rack-cache (v1.5.1 -> v1.6.1)
* unf_ext (v0.0.7.1 -> v0.0.7.2)

I've run all the tests including the regression tests and they all pass.

Note that I've stuck with constraining gds-api-adapters using the
`~> major-version.minor-version` style. We seem to be inconsistent about
how we approach the constraints on gems in the Gemfile, but I'm not going to
worry about that in this PR.