Remove unnecessary test from CountryAndDateQuestionsTest
Now that we have `ErrorHandlingTest`, I think we can be confident that if a
`GdsApi::TimedOutException` is raised then the Smart Answers app will respond
with a 503 HTTP status code.

There are sufficient unit tests in gds-api-adapters' `JsonClientTest` to ensure
that an underlying HTTP timeout exception will be raised as a
`GdsApi::TimedOutException`.

Hence the test in `CountryAndDateQuestionsTest` is unnecessary.