Simplify CountryAndDateQuestionsTest
I think the `exclude_countries` & `include_uk` options for `country_select`
questions are sufficiently well exercised in the `CountrySelectQuestionTest`
unit test.

I don't think that testing this kind of detailed behaviour in an integration
test is very sensible.

Also I find the test very hard to follow i.e. it's not at all obvious why
certain countries should appear or not appear in the select control on the page.

Thus I've removed both the `exclude_countries` & `include_uk` options from the
flow and manually removed United Kingdom & Afghanistan from the list of
countries returned by the Worldwide API.

I've had to change the test to choose "Venezuela" instead of "United Kingdom"
for the 2nd `country_select` question, because the latter is no longer in the
list.