Clarify that WorldLocation.all doesn't return delegations/missions
I hadn't appreciated that the
`GdsApi::TestHelpers::Worldwide#worldwide_api_has_locations` method returns
results with format set to "International delegation" and not "World location".

Also I hadn't appreciated that `WorldLocation.all` only returns results which
have a format of "World location".

Hopefully this change to one of the WorldLocationTest unit tests makes this
clearer.

I came across this when I was trying to understand why the following locations
were not appearing in the `country_select` questions in
`CountryAndDateQuestionsTest`:

* `uk-delegation-to-council-of-europe`
* `uk-delegation-to-organization-for-security-and-co-operation-in-europe`